Update SDK for FRDM-K66 / FRDM-K64 / FRDM-K28

+++ boards/frdmk66f/multiprocessor_examples/erpc_common/erpc_error_handler.cpp
@@ -93,10 +93,10 @@ void erpc_error_handler(erpc_status_t err, uint32_t functionID)
             break;
     }
 
-    /* When error occured on client side. */
+    /* When error occurred on client side. */
     if (functionID != 0)
     {
-        PRINTF("Function id '%u'.", functionID);
+        PRINTF("Function id '%u'.", (unsigned int)functionID);
     }
     PRINTF("\r\n");
